+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/pojo/ApplySupplyCompany.java
@@ -129,7 +129,7 @@ public class ApplySupplyCompany extends CommonBean {
     private String corporateRepresentative;
 
     @ApiModelProperty("注册资金")
-    private Long registeredCapital;
+    private BigDecimal registeredCapital;
 
     @ApiModelProperty("是否禁用")
     private Byte enable;
