+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/manufacturer/vo/ManufacturerReqVo.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -35,7 +36,7 @@ public class ManufacturerReqVo {
     private String legalPerson;
 
     @ApiModelProperty("注册资金")
-    private Long registeredCapital;
+    private BigDecimal registeredCapital;
 
     @ApiModelProperty("公司注册号")
     private String registrationNumber;
